+++ src/asyncio_extensions/_compat.py
@@ -4,10 +4,21 @@ from collections.abc import Awaitable, Callable
 from contextvars import Context
 from typing import ParamSpec, TypedDict, TypeVar
 
+_P = ParamSpec("_P")
+_R = TypeVar("_R")
+
 if sys.version_info >= (3, 12):
+    from inspect import iscoroutinefunction, markcoroutinefunction
     from typing import override
 else:
+    from asyncio import coroutines, iscoroutinefunction
+
     from typing_extensions import override
+
+    def markcoroutinefunction(f: Callable[_P, _R]) -> Callable[_P, _R]:
+        f._is_coroutine = coroutines._is_coroutine  # noqa: SLF001
+        return f
+
 
 if sys.version_info >= (3, 13):
     from typing import TypeIs
@@ -32,10 +43,6 @@ else:
     CreateTaskParams = _CreateTaskParams
 
 
-_P = ParamSpec("_P")
-_R = TypeVar("_R")
-
-
 def is_awaitable(
     func: Callable[_P, _R] | Callable[_P, Awaitable[_R]],
 ) -> TypeIs[Callable[_P, Awaitable[_R]]]:
@@ -46,5 +53,7 @@ __all__ = [
     "CreateTaskParams",
     "TypeIs",
     "is_awaitable",
+    "iscoroutinefunction",
+    "markcoroutinefunction",
     "override",
 ]